Generate speed independent of direction
By generating the x and y components independently within the range [-1,
1], the (x, y) speed will be within a square. Particles traveling in a
more diagonal direction will tend to be faster.

This change will generate the (x, y) speed within the unit circle (still
in the [-1, 1] range for x and for y) so that the total speed is
independent of the direction.